Optimize the integration step
Originally, the two `set(map(tuple, df)))` operations were the rate-limiting step for larger networks. Optimization comes from two places:
1) Using `unique_pairs_df` which is computed once outside of the loop
2) Combining `drop_duplicates()` with a left-join merge in Pandas to avoid expensive conversion from df to set back to df.

I'm not sure if there are any unit tests to ensure that this result is exactly equivalent but it appears to be performing the same overall function. From my private benchmarking, it increased the speed of a particularly large merge from 35 seconds down to 17.